+++ src/wine.h
@@ -1,0 +1,43 @@
+/****************************************************************************/
+//    Copyright (C) 2009 Aali132                                            //
+//    Copyright (C) 2018 quantumpencil                                      //
+//    Copyright (C) 2018 Maxime Bacoux                                      //
+//    Copyright (C) 2020 myst6re                                            //
+//    Copyright (C) 2020 Chris Rizzitello                                   //
+//    Copyright (C) 2020 John Pritchard                                     //
+//    Copyright (C) 2024 Julian Xhokaxhiu                                   //
+//    Copyright (C) 2023 Cosmos                                             //
+//                                                                          //
+//    This file is part of FFNx                                             //
+//                                                                          //
+//    FFNx is free software: you can redistribute it and/or modify          //
+//    it under the terms of the GNU General Public License as published by  //
+//    the Free Software Foundation, either version 3 of the License         //
+//                                                                          //
+//    FFNx is distributed in the hope that it will be useful,               //
+//    but WITHOUT ANY WARRANTY; without even the implied warranty of        //
+//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         //
+//    GNU General Public License for more details.                          //
+/****************************************************************************/
+
+#pragma once
+
+#include <windows.h>
+
+typedef const char* (*WineVersionFunction)();
+const char* GetWineVersion() {
+    HMODULE hModule = LoadLibraryA("ntdll.dll");
+    if (!hModule) {
+        return "Wine is not detected.";
+    }
+
+    WineVersionFunction wine_get_version = (WineVersionFunction)GetProcAddress(hModule, "wine_get_version");
+    if (!wine_get_version) {
+        FreeLibrary(hModule);
+        return "Wine is not detected.";
+    }
+
+    const char* version = wine_get_version();
+    FreeLibrary(hModule);
+    return version;
+}